backend/pyproject.toml
[tool.poetry]
name = "food-discovery-backend"
version = "1.0.0"
description = ""
authors = ["Kacper Łukawski <kacper.lukawski@qdrant.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.100.1"
qdrant-client = "^1.3.2"
python-dotenv = "^1.0"
uvicorn = "^0.23.2"
sentence-transformers = "^2.2.2"
torch = [
    {url="https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.0.1%2Bcpu.cxx11.abi-cp310-cp310-linux_x86_64.whl", markers="python_version=='3.10'"},
    {url="https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.0.1%2Bcpu.cxx11.abi-cp311-cp311-linux_x86_64.whl", markers="python_version=='3.11'"},
]

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
pytest = "^7.4"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
